third_party/WebKit/LayoutTests/fast/dom/DOMException/max-recursion-depth.html

<!doctype html>
<script>

  if (window.testRunner)
    testRunner.dumpAsText();

  var count = 43;
  window.created = 0;

  function define(name, child) {
    var prototype = Object.create(HTMLElement.prototype);
    prototype.createdCallback = function() {
      created++;
      if (created == count)
        document.body.appendChild(document.createTextNode("PASS."))
      document.createElement(child);
    };
    document.registerElement(name, {prototype: prototype});
  }

  var prefix = 'x-';
  for (var i=0; i < count; i++) {
     define(prefix + i, prefix + (i+1));
  }
</script>
<p>
Tests that we can recurse up to 43 times between V8 and Blink.
You should see PASS below, and no stack size exceeded RangeError.
<x-0></x-0>
