third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/DOMException/max-recursion-depth-expected.txt

CONSOLE ERROR: line 18: Uncaught TypeError: document.registerElement is not a function
Tests that we can recurse up to 43 times between V8 and Blink. You should see PASS below, and no stack size exceeded RangeError.
